refactor(glyph): extract whitespace-significant emission to submodule
Keeps formatter.rs and template.rs within the source-file-length budget
by moving the verbatim whitespace-significant element branch and its
tests into template/formatter/whitespace_significant.rs.
